Shift documentation deployment from ReadTheDocs to GitHub Pages
This commit removes the ReadTheDocs configuration file and updates the documentation URL in various locations, including the pyproject.toml file and README.md, to point to GitHub Pages. Additionally, a new GitHub Actions workflow for deploying to GitHub Pages has been added. Documentation will now be hosted on GitHub Pages and updated automatically with each push to the main branch.

Signed-off-by: Jag_k <jag-k@users.noreply.github.com>